+++ src/app/cart/models/cart-item.model.ts
@@ -3,6 +3,7 @@ import { Product } from '../../products/models/product.model';
 export class CartItem {
   constructor(
     public product: Product,
-    public quantity: number = 1
+    public quantity: number = 1,
+    public totalPrice: number = product.price
   ) { }
 }