+++ src/app/products/components/product-list/product-list.component.ts
@@ -9,7 +9,7 @@ import { ProductsService } from '../../services/products.service';
   styleUrls: ['./product-list.component.scss']
 })
 export class ProductListComponent implements OnInit {
-  products: Array<Product>;
+  products: Promise<Array<Product>>;
 
   constructor(
     private productsService: ProductsService,