+++ src/app/shared/shared.module.ts
@@ -2,18 +2,21 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HighlightDirective } from './directives/highlight.directive';
 import { BorderDirective } from './directives/border.directive';
+import { OrderByPipe } from './pipes/order-by.pipe';
 
 @NgModule({
   declarations: [
     HighlightDirective,
-    BorderDirective
+    BorderDirective,
+    OrderByPipe
   ],
   imports: [
     CommonModule
   ],
   exports: [
     HighlightDirective,
-    BorderDirective
+    BorderDirective,
+    OrderByPipe
   ]
 })
 export class SharedModule { }