+++ src/app/shared/pipes/order-by.pipe.ts
@@ -5,18 +5,29 @@ import { Pipe, PipeTransform } from '@angular/core';
 })
 export class OrderByPipe implements PipeTransform {
   transform(value: Array<object>, key: string = null, isAsc: boolean = false): unknown {
+    console.log(value, key, isAsc);
     if (key && key.trim().length) {
       return value.sort((a, b) => {
         const aa = this.getDeepProperty(key, a);
         const bb = this.getDeepProperty(key, b);
-        return isAsc ? aa - bb : bb - aa;
+        if (typeof aa === 'number' && typeof bb === 'number') {
+          return isAsc ? aa - bb : bb - aa;
+        } else {
+          if (aa > bb) {
+            return isAsc ? 1 : -1;
+          } else if (aa < bb) {
+            return isAsc ? -1 : 1;
+          } else {
+            return 0;
+          }
+        }
       });
     }
 
     return value;
   }
 
-  private getDeepProperty(key: string, obj: object) {
+  private getDeepProperty(key: string, obj: object): any {
     const keys = key.split('.');
     return keys.reduce((acc, val) => acc[val], obj);
   }