+++ changelog.md
@@ -41,3 +41,4 @@ Added Border directive
 Fixed task 3 comments
 Added currency pipe to cart-item, cart-list and product components
 Changed getProduct to return Promise; Added async pipe to products-list component
+Added orderBy pipe; Small improvements;